pyproject.toml
[project]
name = "wku-crreservation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4==4.12.3",
    "fastapi>=0.110,<1.0",
    "ics>=0.7.2",
    "pandas==2.2.3",
    "psycopg[binary]==3.2.3",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "selenium==4.25.0",
    "sqlalchemy==2.0.35",
    "uvicorn[standard]>=0.30,<1.0",
    "webdriver-manager==4.0.2",
]
[project.scripts]
crawl = "smartcampus_crawler.crawler:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["smartcampus_crawler"]
